# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ErrorDetail" : {
        "description" : "One field-level validation failure",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "messageKey" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ErrorEnvelope" : {
        "description" : "Uniform error response for every non-2xx API response",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "$ref" : "#/components/schemas/ErrorDetail"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "messageKey" : {
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LocalizedText" : {
        "properties" : {
          "ar" : {
            "type" : "string"
          },
          "en" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SchemaDetail" : {
        "properties" : {
          "claimsDefJson" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "defaultMaxUses" : {
            "format" : "int32",
            "type" : "integer"
          },
          "defaultValidity" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nameI18n" : {
            "$ref" : "#/components/schemas/LocalizedText"
          },
          "sdFields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "status" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apiKeyBearer" : {
        "bearerFormat" : "khk_...",
        "description" : "API key bearer token: Authorization: Bearer khk_<prefix>.<secret>. Scope and actor-kind requirements vary per endpoint — see each operation's description.",
        "scheme" : "bearer",
        "type" : "http"
      },
      "sessionCookie" : {
        "description" : "Console session cookie, set by POST /api/v1/auth/login. Used together with the XSRF-TOKEN cookie/X-XSRF-TOKEN header on state-changing requests.",
        "in" : "cookie",
        "name" : "KHATM_SESSION",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "title" : "Khatm Platform API",
    "version" : "0.1.0-SNAPSHOT"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/schemas/{id}/publish" : {
      "post" : {
        "description" : "Requires the admin scope. DRAFT -> PUBLISHED — the immutability line: a published schema's claim fields can never be mutated again (no general update endpoint exists for a PUBLISHED schema); a mistake found later needs a new version (POST /{id}/versions) instead. A PUBLISHED schema becomes a valid issuance target for POST /api/v1/credentials/issue's schemaCode.",
        "operationId" : "publish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaDetail"
                }
              }
            },
            "description" : "Schema published"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "No valid session or API key"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "Missing the admin scope"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "No schema with this id (KH-SCH-0404)"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "The schema is not DRAFT (KH-SCH-1409)"
          }
        },
        "summary" : "Publish a DRAFT schema",
        "tags" : [ "schema" ]
      }
    },
...
    "/sl/{tenantSlug}/{listCode}" : {
      "get" : {
        "description" : "The public, unauthenticated source of revocation truth for offline verifiers (spec FS-1.3 D2, SAD §6). Returns the list's signed bitstring as a compact JWS (application/jose) — a verifier validates its signature against the tenant's own JWKS, then base64url-decodes and gunzips the `bits` claim to read the per-credential revocation bit at the index named in the credential's `status` claim. The ETag is the list's `version`; a matching If-None-Match returns 304 with no body. Cached for 60s.",
        "operationId" : "getStatusList",
        "parameters" : [ {
          "description" : "the tenant's slug",
          "in" : "path",
          "name" : "tenantSlug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "the status list's code",
          "in" : "path",
          "name" : "listCode",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "If-None-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/jose" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "The signed status-list artifact",
            "headers" : {
              "Cache-Control" : {
                "description" : "max-age=60",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "ETag" : {
                "description" : "The list's version, quoted — weak entity tag for revalidation",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "304" : {
            "content" : {
              "application/jose" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "The client's If-None-Match matched the current version — body omitted"
          },
          "404" : {
            "content" : {
              "application/jose" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "No status list at this tenantSlug/listCode, or the tenantSlug is unknown (KH-TNT-0404/KH-STS-0404)"
          }
        },
        "summary" : "Fetch a signed status-list artifact",
        "tags" : [ "status" ]
      }
    },
    "/t/{tenantSlug}/.well-known/jwks.json" : {
      "get" : {
        "description" : "Public ACTIVE + RETIRING signing keys for the named tenant, no authentication required. Stays available regardless of the tenant's suspension status (spec V4).",
        "operationId" : "jwks",
        "parameters" : [ {
          "description" : "the tenant's slug",
          "in" : "path",
          "name" : "tenantSlug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "The tenant's JWKS"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description" : "No tenant with this slug (KH-TNT-0404)"
          }
        },
        "summary" : "Fetch a tenant's JWKS",
        "tags" : [ "jwks" ]
      }
    }
  },
...
}